refactor: move components to seperate files
